Adjust namespace for moved files
