+++ src/Microsoft.OpenApi.Readers/V2/OpenApiContactDeserializer.cs
@@ -5,6 +5,7 @@
 
 using System;
 using Microsoft.OpenApi.Any;
+using Microsoft.OpenApi.Extensions;
 using Microsoft.OpenApi.Models;
 using Microsoft.OpenApi.Readers.ParseNodes;
