+++ src/Microsoft.OpenApi.Readers/V3/OpenApiResponseDeserializer.cs
@@ -5,6 +5,7 @@
 
 using System.Collections.Generic;
 using Microsoft.OpenApi.Any;
+using Microsoft.OpenApi.Extensions;
 using Microsoft.OpenApi.Models;
 using Microsoft.OpenApi.Readers.ParseNodes;
