+++ test/Microsoft.OpenApi.Tests/Models/OpenApiOperationTests.cs
@@ -6,6 +6,7 @@
 using System;
 using System.Collections.Generic;
 using FluentAssertions;
+using Microsoft.OpenApi.Extensions;
 using Microsoft.OpenApi.Models;
 using Xunit;
 using Xunit.Abstractions;
