+++ test/Microsoft.OpenApi.Tests/Models/OpenApiParameterTests.cs
@@ -5,6 +5,7 @@
 
 using System.Collections.Generic;
 using FluentAssertions;
+using Microsoft.OpenApi.Extensions;
 using Microsoft.OpenApi.Models;
 using Xunit;
 using Xunit.Abstractions;
