+++ test/Microsoft.OpenApi.Tests/Models/OpenApiResponseTests.cs
@@ -4,6 +4,7 @@
 // ------------------------------------------------------------
 
 using FluentAssertions;
+using Microsoft.OpenApi.Extensions;
 using Microsoft.OpenApi.Models;
 using Xunit;
