+++ test/Microsoft.OpenApi.Tests/Models/OpenApiSchemaTests.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using FluentAssertions;
 using Microsoft.OpenApi.Any;
+using Microsoft.OpenApi.Extensions;
 using Microsoft.OpenApi.Models;
 using Xunit;
 using Xunit.Abstractions;
