+++ test/Microsoft.OpenApi.Tests/Models/OpenApiTagTests.cs
@@ -6,6 +6,7 @@
 using System.Collections.Generic;
 using FluentAssertions;
 using Microsoft.OpenApi.Any;
+using Microsoft.OpenApi.Extensions;
 using Microsoft.OpenApi.Models;
 using Xunit;
